requirements.txt
openai
psycopg2-binary
uvicorn
fastapi
pydantic
load_dotenv
python-telegram-bot==12.8
python-dotenv
